+++ frontend/src/components/Home.js
@@ -5,6 +5,8 @@ import { Redirect } from "react-router-dom";
 import { sessionService } from "redux-react-session";
 import axios from "axios";
 import Logo from "../assets/OrangeLogo_outline.png";
+import TwitterLogin from "react-twitter-auth";
+import { getTwitterEntries } from "../actions";
 
 class Home extends Component {
   constructor(props) {
@@ -12,7 +14,10 @@ class Home extends Component {
 
     this.state = {
       currentEntry: "",
-      redirectTo: null
+      redirectTo: null,
+      isAuthenticated: false,
+      user: null,
+      token: ""
     };
   }
 
@@ -41,15 +46,61 @@ class Home extends Component {
     });
   };
 
+  onSuccess = response => {
+    const token = response.headers.get("x-auth-token");
+    response
+      .json()
+      .then(user => {
+        if (token) {
+          this.setState({ isAuthenticated: true, user: user, token: token });
+        }
+      })
+      .then(() => {
+        this.setState({
+          redirectTo: "twitter"
+        });
+      });
+  };
+
+  onFailed = error => {
+    alert(error);
+  };
+
+  logout = () => {
+    this.setState({ isAuthenticated: false, token: "", user: null });
+  };
+
   render() {
     if (this.state.redirectTo === "graph") {
       return <Redirect to="/graph" />;
+    } else if (this.state.redirectTo === "twitter") {
+      return <Redirect to="/settings" />;
     }
+
+    let content = this.state.isAuthenticated ? (
+      <div>
+        <p>Authenticated</p>
+        <div>
+          <button onClick={this.logout} className="button">
+            Log out
+          </button>
+        </div>
+      </div>
+    ) : (
+      <TwitterLogin
+        loginUrl="http://localhost:3000/auth/twitter"
+        onFailure={this.onFailed}
+        onSuccess={this.onSuccess}
+        requestTokenUrl="http://localhost:3000/auth/twitter/reverse"
+        text=""
+      />
+    );
+
     return (
       <section className="hero is-fullheight">
         <div className="hero-head">
           <div className="level is-mobile" id="loginBox">
-            <div className="level-left" />
+            <div className="level-left">{content}</div>
             <div className="level-right">
               <div className="level-item">
                 <button className="button" onClick={this.handleLogout}>
